Added tests, then refactored the impl based on Jesse's suggestions.
