+++ src/ocean_emulators/datasets.py
@@ -1,18 +1,15 @@
-import itertools
 import logging
-import random
 import time
-from collections.abc import Callable
 from concurrent.futures import wait
 from concurrent.futures.thread import ThreadPoolExecutor
-from typing import Any, Self, TypeAlias, final
+from typing import TypeAlias, final
 
 import numpy as np
 import torch
 import xarray as xr
 from einops import rearrange
 from jaxtyping import Float
-from torch.utils.data import BatchSampler, Dataset, Sampler, SubsetRandomSampler
+from torch.utils.data import Dataset
 from xarray_einstats.einops import rearrange as xr_rearrange  # noqa: F401
 
 from ocean_emulators.constants import (
@@ -697,160 +694,6 @@ def concurrent_compute(
             futures.append(executor.submit(load_variable_data, var))
 
     wait(futures)
-
-
-class _SimpleSubsetSampler(Sampler):
-    def __init__(self, indices):
-        super().__init__()
-        self.indices = indices
-
-    def __iter__(self):
-        return iter(self.indices)
-
-    def __len__(self):
-        return len(self.indices)
-
-
-class EquivalenceGroupBatchSampler(Sampler[list[int]]):
-    """Groups indices by dataset membership in ConcatDataset, batches within groups, and optionally shuffles.
-
-    This sampler partitions dataset indices into groups based on their source dataset when using
-    ConcatDataset. It creates batches within each group, then chains them together. When shuffle=True,
-    batches are globally shuffled each epoch to avoid sequential group processing.
-
-    Args:
-        dataset_sizes: List of individual dataset sizes. Groups are created based on dataset boundaries,
-            where each dataset forms its own equivalence group.
-        batch_size: Number of samples per batch
-        shuffle: Whether to shuffle indices within groups and shuffle batches globally
-        drop_last: Whether to drop incomplete batches at the end of each group
-    """
-
-    def __init__(
-        self,
-        dataset_sizes: list[int],
-        batch_size: int,
-        shuffle: bool = True,
-        drop_last: bool = False,
-    ):
-        super().__init__()
-        self.group_size = len(dataset_sizes)
-        self.batch_size = batch_size
-        self.shuffle = shuffle
-        self.drop_last = drop_last
-
-        # Create groups based on cumulative dataset boundaries
-        cumsum = 0
-        self.groups = []
-        for size in dataset_sizes:
-            self.groups.append(list(range(cumsum, cumsum + size)))
-            cumsum += size
-
-    @classmethod
-    def from_datasets(
-        cls,
-        datasets: list["TorchTrainDataset"],
-        group_key: Callable[["TorchTrainDataset"], Any],
-        batch_size: int,
-        shuffle: bool,
-        drop_last: bool,
-    ) -> Self:
-        """Create sampler by grouping datasets using a key function.
-
-        This factory method allows grouping datasets by arbitrary criteria (e.g., resolution,
-        regardless of other parameters like stride). Datasets with the same key are batched together.
-
-        Args:
-            datasets: List of TorchTrainDataset instances to group
-            group_key: Callable that extracts grouping key from a dataset.
-            batch_size: Number of samples per batch
-            shuffle: Whether to shuffle indices within groups and shuffle batches globally
-            drop_last: Whether to drop incomplete batches at the end of each group
-
-        Examples:
-                - lambda ds: (ds._input_src.data.sizes['lat'], ds._input_src.data.sizes['lon'])  # group by resolution
-                - lambda ds: ds._input_src.data.sizes['lat']  # group by latitude size only
-            batch_size: Number of samples per batch
-            shuffle: Whether to shuffle indices within groups and shuffle batches globally
-            drop_last: Whether to drop incomplete batches at the end of each group
-
-        Returns:
-            EquivalenceGroupBatchSampler configured to group by the provided key
-
-        Example:
-            >>> # Group datasets by resolution, allowing different strides to be batched together
-            >>> sampler = EquivalenceGroupBatchSampler.from_datasets(
-            ...     datasets=dataset_list,
-            ...     group_key=lambda ds: (ds._input_src.data.sizes['lat'], ds._input_src.data.sizes['lon']),
-            ...     batch_size=32,
-            ...     shuffle=True,
-            ...     drop_last=True,
-            ... )
-        """
-        from collections import defaultdict
-
-        # Group indices by their key
-        groups: dict[tuple, list[int]] = defaultdict(list)
-
-        cumsum = 0
-        for ds in datasets:
-            key = group_key(ds)
-            # Make key hashable if it isn't already
-            if not isinstance(key, (int, str, tuple)):
-                key = tuple(key) if hasattr(key, "__iter__") else (key,)
-            groups[key].extend(range(cumsum, cumsum + len(ds)))
-            cumsum += len(ds)
-
-        # Convert groups to dataset_sizes format
-        # Sort by key for deterministic ordering across runs
-        sorted_groups = sorted(groups.items(), key=lambda x: str(x[0]))
-
-        # Create instance with computed dataset_sizes
-        instance = cls.__new__(cls)
-        instance.batch_size = batch_size
-        instance.shuffle = shuffle
-        instance.drop_last = drop_last
-        instance.group_size = len(sorted_groups)
-
-        # Store the actual grouped indices (not just sizes)
-        instance.groups = [indices for _, indices in sorted_groups]
-
-        return instance
-
-    def __iter__(self):
-        # Choose sampler based on shuffle setting
-        SubsetSampler = SubsetRandomSampler if self.shuffle else _SimpleSubsetSampler
-
-        # Create batch samplers for each group
-        batch_sampler = itertools.chain(
-            *[
-                BatchSampler(
-                    SubsetSampler(group),
-                    batch_size=self.batch_size,
-                    drop_last=self.drop_last,
-                )
-                for group in self.groups
-            ]
-        )
-
-        if not self.shuffle:
-            # No global shuffle: return batches in sequential group order
-            yield from batch_sampler
-        else:
-            # Shuffle batches globally to avoid sequential group processing
-            # This is regenerated each epoch, giving different orderings
-            all_batches = list(batch_sampler)
-            random.shuffle(all_batches)
-            yield from all_batches
-
-    def __len__(self):
-        """Calculate total number of batches across all groups."""
-        total_batches = 0
-        for group in self.groups:
-            if self.drop_last:
-                total_batches += len(group) // self.batch_size
-            else:
-                total_batches += (len(group) + self.batch_size - 1) // self.batch_size
 
 
 @final
